refactor(ui): extract Graph tab to ui/tabs/graph.rs
Lifts draw_graph_tab plus its seven sub-renderers (traffic chart,
connections sparkline, app distribution, top processes, health
chart, TCP counters, TCP states) into ui/tabs/graph.rs (~575
lines). Detail-panel constants that lived in the same span move
back to mod.rs where the Details renderer can still see them.
