+++ src/ui/tabs/mod.rs
@@ -2,5 +2,6 @@
 //! tab (Overview, Details, Interfaces, Graph, Help) and is invoked
 //! from the top-level `draw()` dispatcher in `ui::mod`.
 
+pub(super) mod graph;
 pub(super) mod help;
 pub(super) mod interfaces;
